% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_julia_output.R
\name{read_julia_output}
\alias{read_julia_output}
\title{Read Julia input data and output data}
\usage{
read_julia_output(juliaout_path, data_path)
}
\arguments{
\item{juliaout_path}{character, path to julia output of fitted
chains, extracted density evaluations, estimated locational
desirability and plots}

\item{data_path}{character, path to data directory for project}
}
\value{
list, list of data.tables
}
\description{
Reads and augments data and output from julia fit
}
\author{
Konstantin Hoffie
}
